requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pluggy==1.0.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.2
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.5.post1
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
